Added correct tests

+++ src/main/java/ru/itsjava/dao/EmailJdbc.java
@@ -2,12 +2,11 @@ package ru.itsjava.dao;
 
 import ru.itsjava.domains.Email;
 
-import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
 public interface EmailJdbc {
-    BigInteger countEmailByAddress(String address);
+    long countEmailByAddress(String address);
 
     Optional<Email> getEmailById(long id);
