+++ src/main/java/ru/itsjava/dao/UserJdbc.java
@@ -2,12 +2,11 @@ package ru.itsjava.dao;
 
 import ru.itsjava.domains.User;
 
-import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
 public interface UserJdbc {
-    BigInteger countUserByName(String fio);
+    long countUserByName(String fio);
 
     Optional<User> getUserById(long id);
